Avoid null ptr dereference on mount failure
With the new multiple mount code, if the mount fails after the nx info
has been attached to the superblock (for instance, because an invalid
option was supplied), then blkdev_put() will be called twice for the
device, once directly and once in apfs_unmap_main_super().

To avoid this, only call blkdev_put() directly before the block device
is assigned to the nx info. Lock the mutex before getting the device, to
avoid inverting the cleanup order, though I don't think that would
matter here.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>