Don't leak a node's freed record space
Right now the module doesn't bother reusing free space from a node until
it's split. This is a serious problem because it's common for the ip
free queue to be banned from ever having more than one node.

For reuse to be possible, whenever a record is removed (or resized),
keep track of the free space: either move the edges of the key/value
data if it's a tail record, or add the new hole to the proper free list.
Then, on allocation, try to use space from the free lists first.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>