Preallocate toc for nodes with fixed record sizes
Currently the Apple fsck gives the following error:

  error: btn: invalid btn_table_space (0, 320), given btn_flags (0x5)

Stan figured out that, when a node has keys and values of a fixed size,
we are expected to preallocate a table of contents that covers the whole
node. So do that.

Weirdly, the node space calculation doesn't seem to take into account
the presence of a footer in root nodes.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>