Avoid npd on dentry hash collisions
Check that none of the names being compared by apfs_keycmp() is NULL
before passing them to strcmp(), to avoid a null pointer dereference.
This was originally never supposed to happen, because keys that reach
this point in the comparison are of the same type; but it does happen
for dentry record insertions because of the hacky way we handle hash
collisions. We still need to properly support that.

[ernesto: added commit message]

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>