Support insertion to the first position of a btree
Record insertions are not usually done in the first position of a node:
the record would go to the previous node instead. So, the case when
there is no previous node needs to be treated separately.

This problem wasn't noticed before because catalog insertions are never
done in the first position, since the first few are always taken by
records for the root directory's parent and for the private directory.
It's for the extent reference tree that this becomes noticeable, and
when old blocks are actually being freed, which was only recently
implemented.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>