Allow multiple operations per transaction
Assume no errors, and split the transaction into nx and volume
structures. The commit decides if it wants to commit everything or
nothing at all.

This doesn't work correctly for extent writes, because something is
making the buffer unmapped before the real commit. The next patch will
deal with this issue.

[ernesto: added commit message]
Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>