Write extent content to disk right away
The last patch broke extent writes, because something is making the
buffer head unmapped before the commit. I failed to figure out the
cause, but the issue can be avoided if the content is written to disk
right away, so do that. It should be noted that this has a performance
cost, and that we could leak the buffers in a short write. In the long
term, it would be better to rewrite everything using bios instead of
buffer heads.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>