Adjust the checks that decide if commit is needed
Extent blocks are not currently counted as part of the transaction, so
it would take too long to reach the 8196 threshold for a commit. It's
important that most transactions do not end because a free queue is
full, otherwise we get a burst of short transactions while the queue is
flushed.

So reduce the buffer limit substantially, and also increase the blocks
allowed in the main queue. For the internal pool, simply check that it
has enough room to map the whole container. Finally, add a check on the
total number of times the transaction was started, to assure regular
commits on file writes even if the metadata remains small, which is now
expected because of longer extents.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>