Allow tail extents to be enlarged on append
Keep the latest extent written in the extent cache, and enlarge it when
possible instead of starting a new one. Always write the dirty cache on
apfs_write_end(); this reduces the value of using a cache, so I should
consider keeping track of dirty inodes inside the transaction, and
writing both the extent and the inode metadata only on commit.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>