Try to keep extent and metadata blocks separate
Our allocation scheme is very naive for now, we just look for the first
available block. We can still allocate large extents this way because we
only write to one file at a time, but the process is interrupted as soon
as a metadata block is needed.

Until I can work on improving allocation, try to avoid this problem by
allocating metadata blocks at the end of the disk.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>