Don't reuse negative dentries on file creation
If we run

  touch aaa
  rm aaa
  touch AAA

on a case-insensitive filesystem, the result will be a file called
"aaa", instead of the expected "AAA". This happens because a negative
dentry for "aaa" still exists in the cache, and since it matches "AAA",
it gets reused for the creation.

The ->d_revalidate() method exists to deal with these situations, so
implement it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>