Set oldest xid in free queues correctly
When a free queue is flushed, the oldest xid reported usually needs to
change. Currently we just increase it by one, which is correct if we
assume that all transactions add at least one record to the free queue.
This assumption breaks down when we run out of space, because that may
trigger a transaction that does nothing but flush the queues.

So set the oldest xid correctly, by checking the first record from the
free queue and using that value.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>