Do proper cleanup of superblock on mount failure
If the main superblock can't be read on mount, we currently leak
references to both the superblock and the module. This makes it
impossible to use rmmod, and triggers a deadlock if the mount is retried
a couple of times.

Once we have a superblock, always perform proper cleanup with
deactivate_locked_super(). The ->put_super() method will not be engaged
unless we've been through apfs_fill_super() already, which sets s_op.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>