Silence compiler warnings in apfs_create_node()
Some compilers complain that bno and oid may be used uninitialized in
apfs_create_node(). This isn't true in practice, because the code path
that triggers the warning will never be run if the function was called
correctly; this is documented by an "ASSERT(false)".

Deal with it anyway, both to avoid the noise during compilation, and to
prevent serious damage if I accidentally misuse this function in the
future.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>