Take exclusive ownership of block device on mount
I've been tracking down a bug that seemed a bit surreal at first. If I
run:

  mount -o vol=1 device dir1
  mount device dir2

the kernel gets livelocked as soon as the second mount tries to map a
single block. But there is no problem at all if I run the entirely
equivalent:

  mount -o vol=1 device dir1
  mount -o vol=0 device dir2

Since I'm not specifying the filesystem type, the kernel first tries to
mount the device using my root filesystem, which is xfs. There is no
"vol" option in xfs, so if it's included this process fails as soon as
options are parsed. If not, xfs will proceed to look at the superblock;
this involves messing with the device, which is already in use since
volume 1 was mounted. The kernel is left in a weird state.

I made this mess when I first tried to follow the mount code from btrfs.
They didn't seem to use FMODE_EXCL for the block device, so I didn't use
it either, even though I thought it was weird and I even marked it with
XXX. In reality they do use it, they just set it later.

So use FMODE_EXCL, but also restructure the apfs mount code so that the
block device is not taken until after we check if it's already in our
list. Otherwise FMODE_EXCL would not allow multiple mounts at all.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>