Plug memory leak on multiple mount of same volume
On mount, a new superblock info structure is allocated before we
actually confirm that it's needed. If it turns out that the volume was
already mounted, the existing sbi will be used instead, and the new one
needs to be freed, so do it.

Kmemleak does report this issue, but you need to run the scan twice for
some reason.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>