Raise minimum supported kernel version to 4.19
Adding write support meant that the module stopped bulding for the 4.14
kernel:

  https://github.com/linux-apfs/linux-apfs-rw/issues/12

The problem is that discard_new_inode() was only added in 4.19. This is
probably fixable, but it might be a bit messy so I'd rather not waste
time with it unless someone actually needs it. Update the readme with
the new range of supported versions.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>