Set all volumes to read-only on transaction abort
When a transaction is aborted, the whole container must be set to
read-only to avoid the risk of corruption. I don't recall why, but at
some point I thought this would be hard to do, so I just set the
read-only flag in the volume the triggered the abort. Do it right.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>