Mark truly new extent buffers as such
The blocks allocated by apfs_get_new_block() are always new, but the
buffer may hold a copy of an existing block. Truly new buffers need to
be marked as such for their edges to get zeroed, but we don't want that
to happen for regular copy-on-write.

A good way to tell both cases apart is to check the uptodate flag, since
I only set that in apfs_write_begin() after reading a buffer.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>