Implement truncation
Add support for truncation, which is required to finally be able to
delete files.

Growing truncation works by leaving hole extents in the file, and more
work will soon be needed to support writes to such extents. Shrinking
truncation has a different problem, which is that I don't know how to
deal with the already written bytes that are left after the new end. I
decided to zero them out immediately (doing copy-on-write of course),
but if the Apple implementation does something different it will be a
problem when the file grows again.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>