Support node deletion in the extent reference tree
Add support for extent reference tree queries to apfs_delete_node().
The nodes are physical just like those of the object map, so it's a
simple change. It's needed immediately because generic/001 from xfstests
exercises this code path.

One thing to note is that extent reference entries are usually deleted
when an orphan inode gets cleaned up, so the EOPNOTSUPP error code from
this function doesn't reach userland. This made the source of the
problem less obvious, so from now on print an error message to the
console as well.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>