Delete xattrs along with their inode
We currently leak an inode's xattrs when the inode is deleted. Implement
a new apfs_delete_all_xattrs() to fix this. Only actually delete inline
xattrs for now, and abort the transaction if we find any dstream xattrs.
This is good enough for dealing with symlinks to short paths, which is
what we need to pass generic/005 in xfstests.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>