Fix allocation from big entries in node free lists
When a new record key or value is allocated inside the slot left behind
by a larger one that was removed, the free space count gets corrupted.
The problem is that a new entry is added for the room remaining in the
slot, which increases the free space count even though that space was
already counted. Fix this.

After the previous commit ff6e5be159ca ("Avoid splitting nodes with a
single record"), this bug started causing failures in generic/001 of
xfstests. This appears to be just a coincidence and not a problem with
that patch.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>