Fix npd on multiple node split
After a node gets split, its query will no longer have the parent field
set. If the record we intend to insert is sill too big to fit in the
resulting node half, a second split will be triggered, and the lack of a
parent in which to update the index records will result in null pointer
dereference.

To prevent this, check if a query's parent is gone before a split and,
if that's the case, refresh it. To be safe, also add a check for the
parent inside apfs_node_split().

This is all necessary to fix crashes in generic/013 of xfstests, but it
doesn't fully solve the problems with splits yet. When inserting a
sufficiently large inline xattr, it's possible that a defragmented node
with a single record will still be unable to receive it. Add a check for
this and return -EOPNOTSUPP, to prevent an infinite loop; but I will
need to deal with it for real at some point.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>