Include rdev xfield in device inodes with null id
The official implementation creates rdev xfields even for devices with
null major and minor numbers. This is reasonable, but my implementation
neglects it, apparently because I thought that device files with null id
were invalid. Fix it.

A corresponding check has been added to the fsck as 1dcc876db945
("apfsck: allow rdev to be zero for device files"). With that in place,
generic/013 from xfstests will serve as a test for this issue.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>