Refresh queries on multiple record replacements
I've been eyeing apfs_update_inode() with suspicion for a while, because
it replaces the same record multiple times in a row, without refreshing
the query. Now I have confirmation that this is a problem, coming from
generic/013 in xfstests.

To get this to work without big changes, refresh the query inside
apfs_btree_replace() if it's needed.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>