Don't attempt CoW on nonexistent extent blocks
The 'to' and 'from' variables in apfs_write_begin() are supposed to
limit which blocks in a page are preread for copy-on-write, so it makes
no sense if 'to' lies beyond the end of the file. The code was not
written with sparse files in mind though, so a large 'pos' could make
'to' wrap around and become huge. When the module attempts to read those
nonexistent blocks, the extents are not found, and so corruption is
reported and the transaction aborts.

This problem was found by looking at the serial output for generic/013
in xfstests, but that test doesn't report it as a failure.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>