Continue fixing build problems on older kernels
Luflosi has reported some new build failures:

  https://github.com/linux-apfs/linux-apfs-rw/issues/12#issuecomment-890317156

One problem is that the UUID_SIZE macro was not defined until 4.13, so
provide a definition of our own for older kernels. Also, while not
strictly necessary, include the uuid header explicitly (right now it
only works because <linux/fs.h> is included in apfs.h before
"apfs_raw.h").

The other problem is that lockdep_assert_held_write() had a different
name before 5.3, and may not even exist in some older kernels. This
macro is just for debugging, so define it as a no-op when not available.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>