Avoid block_invalidatepage() on inode eviction
When an inode is evicted, its pages are truncated and the buffer heads
discarded even if they are part of the current transaction. This leads
to weirdness on commit time, as first mentioned in patch a78c62f6451d
("Write extent content to disk right away"). To try to prevent this, set
->invalidatepage() to a no-op, and hope for the best. It seems to work,
but I really need to take the time to understand the vfs better at some
point.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>