Defer inode metadata writes
Don't write inode metadata (including the cached extent) to the
memory-mapped blocks until the transaction is about to be committed.
This should allow several consecutive metadata changes to be done in a
single operation. Most importantly, a new extent for appended content
will only be written once, in the end. Right now it's rewritten after
each block.

The performance improvement from this is not as big as I had hoped, it's
under 20%. I'm also sure this isn't the right way to approach this,
since I could be using the dirty inode lists from the vfs. So I'll need
to revise this patch at some point, but I'm leaving it for now.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>