Attempt to support mmap()
The general idea behind copy-on-write mmap() is that the pages become
write-protected when a transaction is committed, and any new writes will
fault and call ->page_mkwrite() again. That method must then perform the
new extent allocations, similarly to ->write_begin().

Attempt to implement all of this, and hope for the best.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>