Don't change the inode's sparse flag until commit
When a transaction creates a hole in a file, the in-memory inode is
immediately flagged as "sparse", but the sparse xfield is only created
on commit. This is a problem because another operation within the same
transaction may delete the new hole, leaving the inode flagged but
without the corresponding xfield.

As a simple fix, only flag the inode on commit, if it still has sparse
bytes. Nothing else is done with the in-memory sparse flag, so it's not
a problem that it doesn't match reality during the transaction.

This patch is needed for generic/029 in xfstests.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>