Don't modify blocks on truncation
When a file is truncated in the middle of a block, the bytes that are
left after the end currently get zeroed. This is probably not the right
thing to do, and I intend to fix it properly eventually, if I can
confirm it with the official implementation.

Meanwhile, the ongoing efforts to support mmap() are conflicting with
this makeshift approach. The problem is that the buffer getting zeroed
could already be mapped for writing, so the mmapped block will become
orphaned after the new CoW, and future writes will never become visible
in the transaction.

So do what I think is the right thing, and stop zeroing out blocks on
truncation. Do it only when new writes extend the file again. I still
need to figure out the proper zeroing for multiple-block extents, and
how this all plays along with snapshots.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>