Support the sync family of syscalls
Implement the methods to support all syscalls such as sync() and
fsync(). Do this by just committing the whole transaction, which is not
very performant but should be technically valid. I would like to go back
to this eventually and get it right, but it will probably be tricky and
I would need to understand how btrfs and others handle it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>