Make transaction commit actually effective
Both generic/034 and generic/039 are reporting that transaction commits
don't get properly enforced when triggered by fsync(), but they work
fine when triggered by sync(). I can't claim to understand how any of
this works, but it seems that sync_dirty_buffer() is not sufficient to
flush pages to disk. From looking at the sync() code, I think what's
missing is a call to filemap_write_and_wait(), so add that.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>