Don't append holes until flushing the extent cache
When the extent cache covers the tail of a file, a flush will only
attempt to grow the last extent record. But if the flush was triggered
because we intend to write beyond the end of the file, a new hole extent
may be placed in the tail, and this operation bypasses the cache. So,
make sure that that the hole creation always comes after the flush.

This bug was found by generic/075 of xfstests.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>