Never commit right after making a page writable
I'm hitting a BUG with generic/247 of xfstests, in the line 2212 of
mm/memory.c for kernel v5.3.

On a successful return, ->page_mkwrite() should leave the page locked,
but this isn't always happening. The problem is that a commit is
sometimes triggered, and that undoes all the work of ->page_mkwrite(),
including taking the lock. I've always known that my implementation of
mmap() made no sense if the transaction was committed right away, but I
didn't stop to consider that it would inevitably happen at random.

So, add a new transaction flag, akin to 'force_commit' but with the
opposite purpose: to prevent an immediate commit. Since the transaction
structure already has too many boolean fields, merge them into a single
'state' integer.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>