Advance readdir context position after each emit
I'm running generic/257 of xfstests, and it reports that getdents64()
returns repeated values for d_off. It's been years since I implemented
this stuff, but it would seem that I thought only the final value of the
context position would matter, and I neglected to update it after each
call to dir_emit(). Fix this.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>