Implement a simple ->remount_fs()
Runnning the test generic/294 leaves this message in the console:

  VFS: Busy inodes after unmount of 2dg. Self-destruct in 5 seconds.  Have a nice day...

The problem is that the test remounts the filesystem as read-only, so
all inodes listed in the transaction will never get flushed. Take care
of them before the remount, with a basic ->remount_fs() method. For now
only support read-only remounts.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>