Put a limit on file sizes
I currently get filesystem corruption after running generic/466 from
xfstests. This is a test of writes to huge filesystem offsets, and the
corruption happens because extent lengths wrap around at 2^56 bytes.

According to the documentation, it appears that the intended file size
limit is 2^63, so the proper way to fix this would involve creating a
large number of maximum-length hole extents. I haven't confirmed this
with the official implementation, but either way it seems like a waste
of effort for current disk sizes. Just ban the creation of files that
wouldn't fit in a single extent.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>