Always create dstream xfields for dstream records
Dstream records are created as soon as a file is modified, but the
dstream xfield is only created when the inode record gets updated (at
commit time) and only if the inode isn't empty. So, if we write to a
new file and truncate it to size 0 in the same transaction, a dstream
record will be inserted with no matching xfield, and the fsck will
report corruption.

To fix this, add an i_has_dstream boolean field to the in-memory inode;
when the inode is updated, check this flag instead of the inode size to
decide if the xfield is needed.

This bug doesn't seem to be covered by any of the xfstests I'm currently
running. I find that surprising, given that it's such a simple thing to
reproduce. Maybe the remaining tests will cover it, but in the meantime
I'm adding a new apfs-specific test to prevent a regression.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>