Move inode dstream data into its own struct
Rework functions dealing with inode content so that they no longer work
with the inode itself, and instead work with a new apfs_dstream_info
struct. This struct will be included in the inode info, so this change
makes little difference right now, but the point is to reuse the dstream
code when dealing with large xattrs.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>