Use the common dstream code to read dstream xattrs
Before we actually implement creation and deletion of dstream xattrs,
first convert the existing (read-only) functionality to use the common
code provided by the last patch.

The code is a little hacky in that it uses temporary buffer heads that
are immediately discarded. It could be done better, but I don't want to
put much effort into it because I plan to get rid of buffer heads soon.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>