Deal with huge records that block new insertions
I never implemented a way to deal with huge records that take over a
whole node because I thought such a thing wouldn't happen unless xattrs
were used. I was wrong: generic/013 from xfstests regularly aborts
transactions when it tries to create a large inode record right next to
a large (symlink) xattr record.

The difficulty here is that my code always inserts a new record right
after another, in the same node. The node is split and defragmented if
full, but the two records will always be consecutive. If they overflow
the node by themselves, nothing can be done.

So, implement a new insertion mechanism for this special case: just
create a whole new node and put the new record in it. This is probably
bad in terms of disk usage, but it should be rare outside of xfstests.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>