Set query parent to null on node defragmentation
The header for apfs_node_split() claims the the query parent will always
be set to NULL on success, but this isn't done correctly if the node is
only defragmented. The problem is that apfs_btree_insert() relies on
this behaviour to decide if a split has been attempted yet, and the
current implementation error will make it loop forever.

So fix apfs_node_split() and make it act as documented.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>