Keep query valid on apfs_node_insert() failure
Update the offset of the current query record key if apfs_node_insert()
expands the table of contents. Otherwise, the function may still find
later on that there's no room for the new record, and return ENOSPC
leaving an inconsistent query. A caller that attempts to refresh such a
query would run into trouble.

This bug was discovered with generic/013 of xfstests.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>