Don't nest queries when deleting dstream xattrs
A run of generic/013 of xfstests eventually corrupts a node's metadata,
later damaging the records it contains. The problem is triggered on the
special case that dstream xattr deletion and the dstream truncation it
requires both affect the same node. Since the query that found the xattr
is still in use while the dstream code spins its own queries, the node
structs in the original query will go stale, and eventually write back
that stale metadata to disk.

The problem is of course, that queries must never be nested, so rework
the code to avoid it. I should think of adding some metadata to queries
on debug builds, so assertions can be put in place to prevent this kind
of time-consuming mistakes.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>