Fix SETFLAGS/GETFLAGS ioctls for newer kernels
Luflosi is reporting new build issues:

  https://github.com/linux-apfs/linux-apfs-rw/issues/19

These problems started when I implemented the FS_IOC_SETFLAGS and
FS_IOC_GETFLAGS ioctls. These ioctls have become officially supported by
the vfs recently, and then some of the helpers I was using got removed.
The proper way to implement this in newer kernels is through inode
methods ->apfs_fileattr_get() and ->apfs_fileattr_set().

So do that, and also fix a new instance of the typical build issue with
new namespace parameters. This time I've taken the trouble to build-test
this with a 5.13 kernel, and I even ran xfstests, so there shouldn't be
any more problems, at least not with that version.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>