Fix signature of ->page_mkwrite() for before 4.11
On ->page_mkwrite(), newer kernels need to retrieve the vm_area_struct
pointer from the vm_fault. For older kernels this would not have been
possible, and the vm_area_struct was passed as an argument instead.
Use macros to get all versions to build.

This problem was reported by Luflosi here:

  https://github.com/linux-apfs/linux-apfs-rw/issues/19#issuecomment-918033297

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>