Fix build for kernels without d_instantiate_new()
It seems that d_instantiate_new() was introduced in 4.17 to fix a
problem with lockdep. I won't use lockdep myself in such an old kernel,
and I don't expect users to care. So, to get the module to build with
those kernels, just revert the code to unlock_new_inode/d_instantiate,
as was done before.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>