Actually fix ->page_mkwrite() for before 4.11
Recent commit 7a94546cb7dd ("Fix signature of ->page_mkwrite() for
before 4.11") attempted to support old kernels that didn't keep the
vm_area_struct as a field inside vm_fault. But I didn't realize that the
field was being accessed at several points of apfs_page_mkwrite(), so
the kernel is still not building. Fix this.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>