Don't sync once for each block in the transaction
When a transaction is committed, every dirty block must be synced to
disk before we can write the checkpoint superblock. Right now we enforce
this by calling sync_dirty_buffer() once for each buffer head in the
transaction. This is very slow though, and the only reason I put up with
it for so long is because I find the buffer head interfaces a bit
intimidating. I'm hoping to get rid of them eventually, but that may
take a while.

So, start submitting write requests for all buffers at once and let the
lower layers handle them as they see fit. In my tests, the time to copy
a 1gb file went down by more than 80%. Still not great, but hopefully
good enough.

The problem I now notice though, is that reads are also extremely slow.
I will need to look into this next.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>