Don't commit transactions during ->write_inode()
Since last commit 0a48c4e603c4 ("Re-enable readahead"), ->write_inode()
has started to fire. I had noticed long ago that it wasn't working, but
I didn't know the reason, and it turns out that the problem was the same
that readahead had: the backing_dev_info.

Sadly, this has caused some problems. If ->write_inode() happens to
trigger a transaction commit, the commit will try to manipulate the same
inode and end with a deadlock. To prevent this, avoid commits during
->write_inode() using the existing APFS_NX_TRANS_DEFER_COMMIT
transaction state. Since it is now possible for the transaction to be
in both the DEFER_COMMIT and the COMMITTING states, rework the states to
be flags.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>