Prevent balancing of dirty pages
Since I first set the bdi with commit 0a48c4e603c4 ("Re-enable
readahead"), I've noticed that writes get slower and slower with time
until they freeze entirely. The filesystem then needs to be remounted.
A look at the backtrace tells me that we are getting stuck at
balance_dirty_pages(), which was previously not called at all because
balance_dirty_pages_ratelimited() checks the capabilities of the bdi.

So, change the capabilities of the bdi so that this stuff behaves as it
did before. I can't claim to understand how any of this works, and I'm
probably not doing the right thing, but at least it's only as bad as it
always was.

This patch also reverts the changes from the last one, d97b6b0c39f0
("Use the proper backing_dev_info"). That stuff seemed to make sense,
but messing with the capabilities of the device bdi triggered some
unexpected failures, so I'd rather avoid it. Instead, copy the fields
that matter into the superblock bdi and change the capabilities there.
In the future, I will consider forwarding the congestion check to the
device as well, but right now I don't understand enough to tell if it
matters.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>