Mark inodes as clean after update on commit
Right now my driver doesn't pay much attention to inode dirtiness, we
just keep dirty inodes on a list and flush it on transaction commit. The
vfs still dirties the inodes though; and when I rsync a large filesystem
tree, all dirty inodes seem to get written at once on unmount, and this
avalanche fills the free queues and triggers an abort.

So, explicitly mark the inodes as clean on commit. As usual of late, I
can only hope that the way I'm doing this is correct, but at least I've
tested it and it seems to work.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>