Don't free buffer heads for metadata on commit
Commit 517dd221df8a ("Forcefully free buffer heads on transaction
commit") caused a regression in generic/321 of xfstests (a crash test),
resulting in some metadata corruption. That patch was only intended to
fix a problem with data pages, so don't bother trying to free metadata
pages. That seems to be enough to make this go away.

I'm very unhappy with my transaction engine right now. It's always been
awkward but right now it's become a pile of hacks that I don't really
understand. I want to rewrite it all without buffer heads, but I need to
do plenty of reading before, to avoid making the same mistakes again.
Hopefully I can get to it next.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>