Implement an omap cache for reads
Profiling a full-disk rsync shows that the driver spends around 25% of
the time inside apfs_omap_lookup_block(). So, implement a simple cache
of recently needed mappings and use that for lookups when possible. The
number of slots in the cache (128) wasn't given much consideration, so
fine-tuning would be good in the future.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>