Get the dentry cache to work
Profiling has shown that the module spends too much time inside
apfs_lookup(). I had never really checked if the dentry cache was
working, so naturally that turned out to be the problem: I had always
assumed that the strings inside qstr structures were null-terminated,
but they may actually be a single component in a pathname, terminated by
a forward slash. The result was that we always searched the cache for
full pathnames, which were naturally never found.

To fix this, always pass the filename length from qstr to the unicode
handlers, and make them work without assuming a null-termination.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>