Don't read disk blocks that will be overwritten
Don't call apfs_sb_bread() to work on blocks that have no valuable data
and will just be overwritten. It's wasteful because the driver must
sleep until the unnecessary read is done. Work instead with a new
apfs_getblk(). The result is a 5% performance improvement when rsyncing
a whole volume.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>