Defer all checksum writes until transaction commit
Most checksum calculations are already done during transaction commit,
but there are a few exceptions, mostly in the spaceman code. When a
large file is copied, close to 10% of the time is spent in these useless
recalculations. So, defer all of them instead.

Sadly (and confusingly), I don't see any actual performance improvement
in my benchmarks, but I think it's better to leave this in place anyway.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>